+++ packages/lib/src/index.ts
@@ -5,8 +5,10 @@ import getNodeTextContent from './getNodeTextContent';
 import DarkModeProps, { Theme } from './DarkModeProps';
 import getTheme from './getTheme';
 export { validateChildren } from './validateChildren';
+export { createSyntheticEvent } from './createSyntheticEvent';
 
 export {
+  StoryMeta,
   storybookArgTypes,
   storybookExcludedControlParams,
   IntrinsicElements,